تکمیل مقایسه

+++ WriteBalance.Domain/Entities/PouyaFinancialRecord.cs
@@ -17,8 +17,8 @@ namespace WriteBalance.Domain.Entities
         public int? Arz_Code { get; set; }
 
         // int in bank refah : Moeen_Code , Moeen
-        public int? Moeen_Code { get; set; } 
-        public int? Moeen { get; set; } 
+        public long? Moeen_Code { get; set; } 
+        public long? Moeen { get; set; } 
         public string? Tafzili { get; set; } 
         public string? Code_Arz_Abbr { get; set; } 
         public string? Sharh_Arz { get; set; }
